+++ src/fsharp/service/ServiceStructure.fs
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation.  All Rights Reserved.  See License.txt in the project root for license information.
 
-namespace Microsoft.FSharp.Compiler.SourceCodeServices
+namespace FSharp.Compiler.SourceCodeServices
 
-open Microsoft.FSharp.Compiler.AbstractIL.Internal.Library
-open Microsoft.FSharp.Compiler.Ast
-open Microsoft.FSharp.Compiler
-open Microsoft.FSharp.Compiler.Range
+open FSharp.Compiler.AbstractIL.Internal.Library
+open FSharp.Compiler.Ast
+open FSharp.Compiler
+open FSharp.Compiler.Range
 
 module Structure =
     /// Set of visitor utilities, designed for the express purpose of fetching ranges