+++ tests/FSharp.Compiler.UnitTests/HashIfExpression.fs
@@ -8,12 +8,12 @@ open System.Text
 open NUnit.Framework
 
 open Internal.Utilities.Text.Lexing
-open Microsoft.FSharp.Compiler
-open Microsoft.FSharp.Compiler.Lexer
-open Microsoft.FSharp.Compiler.Lexhelp
-open Microsoft.FSharp.Compiler.ErrorLogger
-open Microsoft.FSharp.Compiler.ErrorLogger
-open Microsoft.FSharp.Compiler.Ast
+open FSharp.Compiler
+open FSharp.Compiler.Lexer
+open FSharp.Compiler.Lexhelp
+open FSharp.Compiler.ErrorLogger
+open FSharp.Compiler.ErrorLogger
+open FSharp.Compiler.Ast
 
 [<TestFixture>]
 type HashIfExpression()     =