+++ vsintegration/src/FSharp.Editor/Formatting/EditorFormattingService.fs
@@ -11,7 +11,7 @@ open Microsoft.CodeAnalysis.Formatting
 open Microsoft.CodeAnalysis.Host.Mef
 open Microsoft.CodeAnalysis.Text
 
-open Microsoft.FSharp.Compiler.SourceCodeServices
+open FSharp.Compiler.SourceCodeServices
 open System.Threading
 open System.Windows.Forms