+++ vsintegration/tests/UnitTests/LegacyLanguageService/Tests.LanguageService.General.fs
@@ -7,8 +7,8 @@ open System
 open System.IO
 open System.Reflection
 open System.Runtime.InteropServices
-open Microsoft.FSharp.Compiler
-open Microsoft.FSharp.Compiler.SourceCodeServices
+open FSharp.Compiler
+open FSharp.Compiler.SourceCodeServices
 open Microsoft.VisualStudio.FSharp.LanguageService
 open Salsa.Salsa
 open Salsa