+++ fcs/RELEASE_NOTES.md
@@ -458,7 +458,7 @@
 * Replace GetSymbolAtLocationAlternate by GetSymbolUseAtLocation
 
 #### 0.0.40 - 
-* Fixed #86 - Expose Microsoft.FSharp.Compiler.Interactive.Shell.Settings.fsi
+* Fixed #86 - Expose FSharp.Compiler.Interactive.Shell.Settings.fsi
 * Fixed #99 - Add IsNamespace property to FSharpEntity
 
 #### 0.0.39 -